+++ src/wad/archive.rs
@@ -10,9 +10,9 @@ use super::util::{wad_type_from_info, read_binary};
 
 pub struct Archive {
     file: File,
-    index_map: HashMap<WadName, uint>,
+    index_map: HashMap<WadName, usize>,
     lumps: Vec<LumpInfo>,
-    levels: Vec<uint>,
+    levels: Vec<usize>,
     meta: WadMetadata,
 }
 
@@ -36,7 +36,7 @@ impl Archive {
         };
 
         // Read lump info.
-        let mut lumps = Vec::with_capacity(header.num_lumps as uint);
+        let mut lumps = Vec::with_capacity(header.num_lumps as usize);
         let mut levels = Vec::with_capacity(32);
         let mut index_map = HashMap::new();
 
@@ -48,11 +48,11 @@ impl Archive {
             index_map.insert(fileinfo.name, lumps.len());
             lumps.push(LumpInfo { name: fileinfo.name,
                                   offset: fileinfo.file_pos as i64,
-                                  size: fileinfo.size as uint });
+                                  size: fileinfo.size as usize });
 
             if fileinfo.name == b"THINGS\0\0".to_wad_name() {
                 assert!(i_lump > 0);
-                levels.push((i_lump - 1) as uint);
+                levels.push((i_lump - 1) as usize);
             }
         }
 
@@ -68,27 +68,27 @@ impl Archive {
             levels: levels })
     }
 
-    pub fn num_levels(&self) -> uint { self.levels.len() }
+    pub fn num_levels(&self) -> usize { self.levels.len() }
 
-    pub fn get_level_lump_index(&self, level_index: uint) -> uint {
+    pub fn get_level_lump_index(&self, level_index: usize) -> usize {
         self.levels[level_index]
     }
 
-    pub fn get_level_name(&self, level_index: uint) -> &WadName {
+    pub fn get_level_name(&self, level_index: usize) -> &WadName {
         self.get_lump_name(self.levels[level_index])
     }
 
-    pub fn num_lumps(&self) -> uint { self.lumps.len() }
+    pub fn num_lumps(&self) -> usize { self.lumps.len() }
 
-    pub fn get_lump_index(&self, name: &WadName) -> Option<uint> {
+    pub fn get_lump_index(&self, name: &WadName) -> Option<usize> {
         self.index_map.get(name).map(|x| *x)
     }
 
-    pub fn get_lump_name(&self, lump_index: uint) -> &WadName {
+    pub fn get_lump_name(&self, lump_index: usize) -> &WadName {
         &self.lumps[lump_index].name
     }
 
-    pub fn is_virtual_lump(&self, lump_index: uint) -> bool {
+    pub fn is_virtual_lump(&self, lump_index: usize) -> bool {
         self.lumps[lump_index].size == 0
     }
 
@@ -98,7 +98,7 @@ impl Archive {
         self.read_lump(index)
     }
 
-    pub fn read_lump<T: Copy>(&mut self, index: uint) -> Vec<T> {
+    pub fn read_lump<T: Copy>(&mut self, index: usize) -> Vec<T> {
         let info = self.lumps[index];
         assert!(info.size > 0);
         assert!(info.size % mem::size_of::<T>() == 0);
@@ -114,7 +114,7 @@ impl Archive {
         buf
     }
 
-    pub fn read_lump_single<T: Copy>(&mut self, index: uint) -> T {
+    pub fn read_lump_single<T: Copy>(&mut self, index: usize) -> T {
         let info = self.lumps[index];
         assert!(info.size == mem::size_of::<T>());
         self.file.seek(info.offset, SeekSet).unwrap();
@@ -124,9 +124,9 @@ impl Archive {
     pub fn get_metadata(&self) -> &WadMetadata { &self.meta }
 }
 
-#[deriving(Copy)]
+#[derive(Copy)]
 struct LumpInfo {
-    name   : WadName,
-    offset : i64,
-    size   : uint,
+    name  : WadName,
+    offset: i64,
+    size  : usize,
 }
